% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hwd_helpers.R
\name{PLLJoinedResultsForTasks}
\alias{PLLJoinedResultsForTasks}
\title{Retrieve the results of a list of tasks and join the results into a single list object.
This requires the result of each task to be a list object.}
\usage{
PLLJoinedResultsForTasks(tasks, join.attributes = NA)
}
\arguments{
\item{tasks}{A list of PLLTask objects}

\item{join.attributes}{A list of strings specifying which keys to join}
}
\description{
Retrieve the results of a list of tasks and join the results into a single list object.
This requires the result of each task to be a list object.
}

