% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hwd_helpers.R
\name{PLLRemoteTasksWithDefinition}
\alias{PLLRemoteTasksWithDefinition}
\title{Retrieve task objects that exist in the R-Parallel service.}
\usage{
PLLRemoteTasksWithDefinition(definition.identifier)
}
\arguments{
\item{definition.identifier}{The identifier of a function definition, as given in the R-Parallel web interface}
}
\value{
A list of PLLRemoteTask objects for the given definition identifier
}
\description{
Retrieve task objects that exist in the R-Parallel service.
}
\examples{
PLLRemoteTasksWithDefinition("29151F65-18AE-45E5-BD51-6D7E319BB9A8")
}

