% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hwd_helpers.R
\name{PLLResultsForTasks}
\alias{PLLResultsForTasks}
\title{Retrieve the results of a list of tasks.}
\usage{
PLLResultsForTasks(tasks)
}
\arguments{
\item{tasks}{A list of PLLTask objects}
}
\value{
A list of the results of each of the PLLTask objects
}
\description{
Retrieve the results of a list of tasks.
}

