% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rparallel_api.R
\name{PLLSetAccessToken}
\alias{PLLSetAccessToken}
\title{Set an HTTP Access Token for use in requests to the R-Parallel service.
By default, this parameter is read from the RPARALLEL_PAT environment variable, but can be overwritten using this function.}
\usage{
PLLSetAccessToken(accessToken)
}
\arguments{
\item{accessToken}{Your Private Access Token (PAT), which can be found under account details at https://r-parallel.herokuapp.com/}
}
\description{
Set an HTTP Access Token for use in requests to the R-Parallel service.
By default, this parameter is read from the RPARALLEL_PAT environment variable, but can be overwritten using this function.
}
\examples{
PLLSetAccessToken("bd53d890b16773848fff3d226a761ed4") # This is an example key. Your key will be different.
}

